backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==14.0
python-dotenv==1.0.1
motor==3.6.0
google-genai==1.5.0
